Polish delete functionality

+++ posts/views.py
@@ -12,10 +12,16 @@ def show_posts(request, post_id=0):
     # Get a list of the nth most recent posts from the database
     # Build the basic structure of the page
     # In the main tag, include those posts formatted
+    user = None
+#    if request.user.is_authenticated and request.user.groups.exists():
+#        user = {
+#            "user_id": request.user.pk,
+#            "groups": request.user.groups.all()[0].name
+#        }
 
     if post_id == 0:
         blogposts = BlogPost.objects.order_by('-posted_at')
-        return render(request, 'posts/posts.html', {'posts': blogposts})
+        return render(request, 'posts/posts.html', {'posts': blogposts, 'allowed_groups': ['admin', 'super_admin']})
     try:
         blogpost = BlogPost.objects.get(pk=post_id)
         return render(request, 'posts/post.html', {'blogpost': blogpost})
@@ -42,14 +48,28 @@ def create_post(request):
     return render(request, 'posts/create_post.html')
 
 @login_required(login_url='login')
-@allowed_groups(allowed_roles=['admin', 'super_admin'])
-def delete_post(request, post_id):
+#@allowed_groups(allowed_roles=['admin', 'super_admin'])
+def delete_post(request, post_id):    
     try:
         blogpost = BlogPost.objects.get(pk=post_id) 
-        blogpost.delete()
-        return redirect('posts')
     except:
         return HttpResponse("El post no existe") 
+    permission = False
+    group = None
+
+    if request.user.groups.exists():
+        group = request.user.groups.all()[0].name
+        if group in ['admin', 'super_admin']:
+            permission = True
+
+    if blogpost.author_id.pk == request.user.pk:
+        permission = True
+        
+    if permission:
+        blogpost.delete()
+        return redirect('posts')
+    else:
+        HttpResponse("No tienes permiso, puto")
 
 def update_post(request):
     # Get a post ID
